Filter projects by subjects left & redirect status
Previously I was using the API call to project.completedness, however
this may not be a good metric for deciding which projects to promote as
for some projects which have lots of data, having a high completedness
score does not neccessarily equate to having little data left for
classifications.

For now, any live projects which still have more than 500 subjects for
classification are included in the list of projects to promote.

Additionally only include projects which are not built on the old
framework (i.e. not panoptes). So if they have a redirect link, do not
include them. This is becuase calling the data from these projects will
be tricky and make implementation of systems for evaluation difficult.