Send random project tweet with & without image
We want to test the effectiveness of tweeting with an image and with and
without a project description. The bot now has functions to do all these
options.

In this commit, a short or long message is randomly chosen and it posts
both a tweet with and without an image. Need to randomise the
presence/absence of an image too.